CardContext와 ListContext 통합
